remove unwanted parsing from BuildDelay

+++ pkg/service/record/record.go
@@ -234,7 +234,7 @@ func (r *Recorder) Start(ctx context.Context, reRecord bool) error {
 func (r *Recorder) Instrument(ctx context.Context) (uint64, error) {
 	var stopReason string
 	// setting up the environment for recording
-	appID, err := r.instrumentation.Setup(ctx, r.config.Command, models.SetupOptions{Container: r.config.ContainerName, DockerNetwork: r.config.NetworkName, DockerDelay: r.config.BuildDelay})
+	appID, err := r.instrumentation.Setup(ctx, r.config.Command, models.SetupOptions{Container: r.config.ContainerName, DockerNetwork: r.config.NetworkName, DockerDelay: uint64(r.config.BuildDelay)})
 	if err != nil {
 		stopReason = "failed setting up the environment"
 		utils.LogError(r.logger, err, stopReason)
